resume_parser/lambda-resume-parser/lambda/requirements.txt
boto3==1.34.34
Pillow==10.2.0
PyPDF2==3.0.1
requests==2.31.0
google-generativeai==0.3.2 